requirements.txt
matplotlib
requests
python-dotenv
sqlalchemy
